Archery ego for heavy body armour
This has no relation to the old glove ego that was replaced with hurling.
Appearing on scale, chain, and plate, this ego halves encumbrance for the
purposes of ranged combat penalty. You can play a heavy armour archer once
again, but at a cost (having to find and enchant the relevant ego). Chance
will be adjusted in a later commit bc several other body egos are coming.