New strategic debuff: Ostracism
Intended as the counterpart to malmutate/ill-omen for orbs of winter, this
effect temporarily caps the player's effective piety, restoring it only
through XP gained. This is a stacking effect, which caps effectively piety
at '200 - [ostracism amount]' (so if you had 165 piety and 50 ostracism,
you would count as 115 piety for most purposes and not have access to
whatever your god would offer at 5*). Piety stars with are 'inaccessible'
due to ostracism are indicated by magenta Xs on the piety bar.

This plays considerably better with several gods that have 'unusual' piety
mechanics than outright losing piety does: Ru sacrifices are not undone
(but access to abilities is temporarily limited), Usk piety can still
rise to a certain level, Ash can give smaller stat boosts without
interacting awkwardly with already-cused gear, etc.

(In fact, I also plan to use this to replace the 'lose all piety' trove
cost in a future commit, since it allows those troves to operate better
with multiple gods that are currently vetoed.)

Internally, most checks to you.piety should now use you.piety() (which will
factor in the ostracism effect). you.raw_piety is still used for anything
which interacts with piety directly (ie: all gaining/losing of real piety,
as well as several gifting schedule mechanics that are safer to not
interact with this debuff.)

Gozag and Xom are unaffected by this effect. (It would have been possible
for this to prevent Xom from being in a good mood, but I know regret-index
wants to eventually make adjustments to how Xom mood works altogether, and
she was in favor of Xom not interacting with it for now. Gozag just doesn't
really have piety at all, though, so it feels consistent for them not to
care.)

--------------

Ostracize tile and vfx made by regret-index, using ontoclasm's Burn Books
icon and Denzi's generic / ecumenical altar tile.