Experimental Zot:5 change: Boundless Tesseracts
While the Zot:5 endvault is an iconic and often tense challenge, there have
been some repeated sentiments over the years that the best (or, at least,
a compelling) way to handle it is through a slow and tedious process:
staying a safe distance away and shouting for an incredibly long time to
very slowly lure monsters out of the vault where they can be pulled back to
the stairs or otherwise dealt with individually.

This commit attempts to address this issue by applying forward time
pressure in a relatively transparent and urgent way (while making pulling
things individually back to the stairs a little less guaranteed to be
safe). Each lung of the endvault now contains a Boundless Tesseract, which
is essentially a form of monster generator. So long as they are active,
random floor-appropriate monsters will periodically spawn across the floor.
This is slow at first, but steadily increases in speed the longer the
tesseracts remain alive. This should make ultra-slow strategies inadequate
and push players to be a little more aggressive with making their way into
the vault (while still being gentle enough initially that it does not
heavily disrupt normal strategies).

When tesseracts first 'wake up' (typically from nearby noise), they will
reveal and announce themselves to the player. Killing either tesseract will
also destroy the other (so the player only needs to enter *a* lung, not
both of them.). Finally, the tesseracts come with a 'pool' of XP to help
make their spawns feel less unpleasant while still removing any farming
incentives: the monsters spawned by tesseracts initially give XP normally
at first, but eventually become unrewarding if the player is slow enough.
Killing the tesseract rewards all remaining XP in the pool, so there is no
reason to delay doing so. This XP should hopefully be large enough that
average play will not run into many (or any) unrewarding monsters.

The spawn rate is very provisional at the moment, pending player testing.

(Tile credit goes to regret-index)