Make undead players more vulnerable against doom/banes
In an inverse mirror of the undead having immunity to mutations, make them
take longer to work off doom and banes. (In keeping with the long tradition
of undead having additional vulnerabilities to contrast their immunities).

Poltergeists and revenants have both proven to be quite strong, so I'm sure
they can take this. I actually considered mummies *not* having this (there
are 'sacred' overtones to their flavor which one could reasonably use as
justification to lack this vulnerability), but wasn't sure if the single
exception would be unintuitive or not.

(Death form can also handle this, though I have some hesitations about
vampire form - but having all undead weaknesses was something I laid out
from the start, so this is at least consistent.)

It's possible that 'recover twice as slowly' is a bit aggressive, but I can
tune that down later if it turns out to be too much.