Revise trove contents into new themed locales
The actualy rewards within Trove are mostly severely focused and uneven,
due to all but a few vaults focusing on very singular item categories- all
weapons, all armour, all jewellery, various potions, or various scrolls,
alongside a small number mixing spellbooks and magic staves. Even the ones
that can contain multiple categories can struggle due to the average
quality of placing only a few of some categories (like weapons) being
relatively low. This makes it very difficult to evaluate what the average
quality of troves should be in order to properly set the costs to enter
them. They are also mostly highly-forgettable boxes of items all in a
singular pile, without the decorations introduced throughout much of the
rest of the game, nor any particular themes, which is a somewhat lacking
state of affairs for one of the very few noncombat zones in the game
(compare to e.g. the extreme number of Temple layouts or Bazaar
customers added in e151113.)

This commit revamps a decent number of trove layouts and grants them new
themes plus names dedicated to spell namesakes and unrands, a space which
has made Wizlabs be one of the most memorable and beloved set of portals
in the game. (It also gets in a handful of names that would be very
difficult to make Wizlabs for with the current sprawl of Wizlabs without
substantially more lategame monster options; this isn't meant to be a hard
cap on returning to any of those eventually, and to those ends there is
a single overlap between troves and wizlabs already present). They also
have a rather large set of lua functions distributing their themes across
a variety of different item categories then tailored to species limits,
god restrictions, and player skills, which should make it so every trove
has a relatively more stable floor and ceiling of item quality. (They also
have less items overall- a hard focus on 12 items compared to the prior
16- between said tailoring of average quality and since the trove costs
revamp in 0288ec2 allows recklessly picking up troves on sight in
exchange for significant downsides.)

It'd take an excessively long amount of words to outline just exactly
what each new trove may contain, but to list and summarize each of them
in brief as an introductory teaser:

 * Ozocubu's Refridgerator:
  - freezing weapons, rC+ gear, and Ice Magic
 * The Storm Queen's Palace Crash Site:
  - electric weapons, rElec gear and shields, and Air Magic
 * Leda's Sunken Stockpile
  - heavy weapons, *Slow or ponderous gear, and Earth Magic
 * The Name-Rending Infernalists' Reservoir
  - flaming weapons, rF or Harm gear, and Fire Magic
 * Eringya's Secret Bog
  - spectral weapons, rPois gear, and Alchemy or Summoning
 * Rutra's Hidden Sanctum
  - venomous weapons, Stealth or +Inv gear, and Hexes
 * The Dread Knight's Derelict Charpel
  - draining or reaping weapons, Will+ or +Slay gear, and manuals
 * The First Justicar's Armoury
  - holy wrath weapons, rN gear, and regeneration amulets
 * The Octopus King's Forgotten Garden
  - enough jewellery and armour to exactly match octopode slots,
 * The Lost Hoard of Erebora
  - dragon scales, talismans, hat randarts, and elemental magic
 * a devouring Rift(!)
  - distortion and chaos weapons, random armour, and Translocations
 * an orange crystal hatchery(?)
  - antimagic weapons, Int or Will+ gear, Earth Magic and Summoning

If it turns out the overall item quality ends up feeling too high relative
to rolling troves or not in a game, the total item count can be reduced a
little, or the various consumables keyed to each of these troves can
replace some of the equipment. As is, I'd like to assess how well the
trove cost changes work out and how well the item mixes plus
skill-weighting feel in practice first. There are also still spaces for
conjoining trove themes, if other people want to submit additions to this
trove set in the future, other spells and unrands help pin down some
currently-absent themes, or more item types are added overall.

(This commit also revises the flavour of the Robe of Night to not mention
a real-world goddess in the context of Crawl's highly established
pantheon, adds a proper tilerim to the storm bow, and also adds a new
decorative floor option of inaccessible alcoholic drinks with a tile
based off of CC0 assets by Jetrel.)