Don't ask which item to remove when all are identical (Back-form)
This behavior (ie: not prompting to pick a ring on ring swap when both are
identical) was seemingly lost in 2dbd7c1 and now it is back again in an
mildly enhanced form. (It should now work on any number of any type of
equipment, provided that all swap candidates are identical to each other.
So formicids with 2 identical gloves get the same treatment.)