Add a Sentinel's Mark icon, display it on players by default
While there is technically a force_more involved with alarm traps since
0644d94 and a unique flash over the player for the spell since 82142fb, it
is possible to miss both in the walls of animations, tiles, and text one
can get on a turn, and it's one of the most relevant debuffs one can
otherwise miss. With this new icon being enabled by default in the
tile_player_status_icons option, it will hopefully be yet more difficult
for players to miss it in a relatively non-obtrusive fashion.