Make Zonguldrok's reaping count as evil (#4655).
Since it's not a weapon, but armour that adds the reaping brand to
whatever attack you're using, manually trigger the evil god conduct for
it, just as we do with Damnation. But do this after all other brand
checks, since it effectively gives your weapon the reaping brand on top
of whatever brand it has.

NOTE: Regarding brands with manual conduct triggers, one-shotting a
monster with Damnation, Zonguldrok, or a chaos effect that doesn't
resolve to a brand that works on dead monsters will not trigger the
appropriate conduct when it should, but that's a separate issue, and
this change still brings us closer to the correct behaviour. Also,
there's no warning prompt about potential penance when wielding
Zonguldrok, because all the warning prompt code is written with
the assumption that only weapons can trigger conducts, not armour,
but that's also a separate issue.