Fix Ozocubu's trove failing to place (#4118)
RypoFalem is correct that the shield egos shouldn't specify two weights
for the same item, which entirely fails to place. To try and get some
of the original intent of making the shields rarer, most of the other
other items in its slot have been increased in total weight. Fixes #4118.