Remove disable-autopickup toggle from 'd'rop menu (Ge0FF)
And automatically disable autopickup of dropped jewellery types (since that
will almost always be the desired behavior since the inventory change).

Technically there were still a few niche cases where this toggle could be
useful, but it was argued that they weren't worth keeping this feature
around for and I'm inclined to agree.