Give a more helpful message when the player is trapped behind allies
The situation where a player cannot swap with their allies is a rare one,
but often confusing when it happens (usually due to a pileup of zombies
that cannot cross deep water and a player who can). Try to give a more
specific message, along with a helpful suggestion.

(While it ordinarily would not be ideal to give the player advice instead
of solving a problem directly, I don't think it would be appropriate for
swapping to behave differently in this specific scenario than all others
- eg: by blinking the monster away or displacing the entire crowd
backwards - especially when it's easy to miss that one has gotten into
this situation in the first place.)