Change 'i' menu hotkey behavior yet again (Acrobat)
Instead of 'first match below the cursor' do 'first match within the
current sub-section in which the cursor is located'. (ie: if the player is
hovering a potion, (c) will first select a potion on (c), if one exists,
regardless of whether the cursor is below it.

I apologize for how unstable the interface has been these past few days,
but I hope we're narrowing in on a stable final version of everything now.