Tweak Tiny Tomes
The Tiny Tomes bookshop sells spellbooks with only one spell
in them. With the advent of parchments, there's not really much
of a distinction between those and a spellbook with one spell,
except that the spellbook requires more keypresses to figure out
what the hell is in it.

Rather than completely cut the entire shop, since people seem to
like it, instead it now sells the second-tiniest spellbooks, with
two spells each.

Also, lower the weighting on two of my spellbook shops, since their
frequency was stepping on part of the purpose of parchments.