New orb: Stardust
This orb conjures a group of shooting stars the first time its holder pays
MP while a hostile monster is in sight. This can be from casting spells,
but also many god abilities.

Shooting stars behave a bit like slow foxfires, but push enemies back on
hit and can be freely shot through (so as to not interfere with conjurers
when they repeatedly appear while blasting). The number of shooting stars
created is equal to the MP that was paid at once, and their damage scales
with evocations.

The orb only recharges once its wearer's MP refills to max and they spend
a few turns with no more enemies in sight (which is the closest to 'once
per battle, at the start of battle' that I can think of).

Hopefully this provides an interesting combination of additional damage,
light cover, and disruption for further blasting on different characters.

(Rather than attempt any logic in terms of 'MP cost' for monster spells
and abilities, the number of stars an enemy holding one makes is purely
based on their HD and happens on any spell/spell-based-ability usage.)

-------------------

Tile work by regret-index, using Denzi's wretched star tile, ontoclasm's
Foxfire tile, and Sastreii's starflower tile.